C4-38 Fix (#129)
* C4-38 fix

* remove pdb

* C4-38 fix datastore if statement

* C4-38 address small point
