+++ snovault/elasticsearch/cached_views.py
@@ -19,7 +19,7 @@ from ..resource_views import (
     item_view_expand
 )
 from ..indexing_views import item_index_data
-from ..util import debug_log
+from ..util import debug_log, filter_embedded
 
 
 def includeme(config):
@@ -34,36 +34,6 @@ def cached_view_embedded(context, request):
     if allowed.isdisjoint(request.effective_principals):
         raise HTTPForbidden()
     return filter_embedded(source['embedded'], request.effective_principals)
-
-
-_skip_fields = ['@type', 'principals_allowed']
-def filter_embedded(embedded, effective_principals):
-    """
-    Filter the embedded items by principals_allowed, replacing them with
-    a 'no view allowed' error message if the effective principals on the
-    request are disjointed
-    """
-    _skip_fields = ['@type', 'principals_allowed']
-    # handle dictionary
-    if isinstance(embedded, dict):
-        if 'principals_allowed' in embedded.keys():
-            obj_princ = embedded.get('principals_allowed')
-            allowed = set(obj_princ['view'])
-            if allowed.isdisjoint(effective_principals):
-                embedded = {'error': 'no view permissions'}
-                return embedded
-
-        for name, obj in embedded.items():
-            if isinstance(obj, (dict, list)) and name not in _skip_fields:
-                embedded[name] = filter_embedded(obj, effective_principals)
-
-    # handle array
-    elif isinstance(embedded, list):
-        for idx, item in enumerate(embedded):
-            embedded[idx] = filter_embedded(item, effective_principals)
-
-    # default just return the sucker
-    return embedded
 
 
 @view_config(context=ICachedItem, permission='view', request_method='GET',
